suspect_detection/requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
